requirements.txt
numpy~=1.26.4
httpx~=0.27.2
nonebot2~=2.3.2
nonebot-adapter-onebot~=2.4.4
opencv-python-headless~=4.10.0.84
openvino~=2024.3.0
h2~=4.1.0
fastapi~=0.115.0
uvicorn~=0.31.0
websockets~=13.1